Refactor server.js file and controller references, implement check for jawsdb in heroku
