+++ controllers/api/index.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const eventRoutes = require('./eventRoutes');
+const eventRoutes = require('./event-routes');
 
 router.use('/events', eventRoutes);
