+++ controllers/index.js
@@ -1,7 +1,7 @@
 const router = require('express').Router();
 
 const apiRoutes = require('./api');
-const homeRoutes = require('./homeRoutes');
+const homeRoutes = require('./home-routes');
 
 router.use('/', homeRoutes);
 router.use('/api', apiRoutes);
