+++ package.json
@@ -18,7 +18,6 @@
         "express-openid-connect": "^2.17.1",
         "express-session": "^1.18.0",
         "handlebars": "^4.7.8",
-        "hbs": "^4.2.0",
         "moment-timezone": "^0.5.45",
         "mysql2": "^3.9.2",
         "nodejs-geolocation": "^2.3.0",
